requirements.txt
azure-identity==1.16.1
azure-search-documents==11.6.0b4
azure-storage-blob==12.21.0
environs==11.0.0
fastapi==0.111.0
gunicorn==22.0.0
loguru==0.7.2
openai==1.33.0
pymupdf==1.24.9
python-dotenv==1.0.1
uvicorn[standard]==0.30.1
